Add DifficultyBadge component and integrate into GameTopStrip for improved difficulty display
