+++ SudokuApp/components/GameTopStrip.js
@@ -3,6 +3,7 @@ import { View, StyleSheet, Text } from 'react-native';
 import GameTimer from './GameTimer';
 import ScoreDisplay from './ScoreDisplay';
 import { useGameContext } from '../contexts/GameContext';
+import DifficultyBadge from './DifficultyBadge';
 
 /**
  * Component that displays the top strip with:
@@ -12,40 +13,8 @@ import { useGameContext } from '../contexts/GameContext';
  * Width matches the header for visual consistency
  */
 const GameTopStrip = ({ style }) => {
-  const { theme, difficulty } = useGameContext();
-  
-  // Use default from GameContext.js initial state if difficulty is not available
-  const DEFAULT_DIFFICULTY = 'medium';
-  
-  // Get the badge color from theme based on difficulty
-  const getBadgeColor = () => {
-    // Define fallback colors in case theme colors aren't available yet
-    const fallbackColors = {
-      'easy': '#d4edda',    // Green for easy
-      'medium': '#ffeeba',  // Yellow for medium
-      'hard': '#f8d7da',    // Pink for hard
-      'expert': '#f8d7da'   // Pink for expert
-    };
-    
-    // Use default if difficulty is not specified
-    const difficultyLevel = difficulty || DEFAULT_DIFFICULTY;
-    
-    // Check if theme and theme.colors and theme.colors.difficulty exist
-    if (theme?.colors?.difficulty) {
-      return theme.colors.difficulty[difficultyLevel] || theme.colors.difficulty[DEFAULT_DIFFICULTY];
-    }
-    
-    // Fallback to hardcoded colors if theme isn't fully loaded
-    return fallbackColors[difficultyLevel] || fallbackColors[DEFAULT_DIFFICULTY];
-  };
-  
-  // Get difficulty label with proper capitalization
-  const getDifficultyLabel = () => {
-    // Use default if difficulty is not specified
-    const difficultyLevel = difficulty || DEFAULT_DIFFICULTY;
-    return difficultyLevel.charAt(0).toUpperCase() + difficultyLevel.slice(1);
-  };
-  
+  const { theme } = useGameContext();
+
   return (
     <View style={[styles.container, { backgroundColor: theme?.colors?.background || '#f8f8f8' }, style]}>
       {/* Left section: Score */}
@@ -55,12 +24,7 @@ const GameTopStrip = ({ style }) => {
 
       {/* Center section: Difficulty level */}
       <View style={styles.centerSection}>
-        <Text style={styles.levelLabel}>LEVEL</Text>
-        <View style={[styles.levelBadge, { backgroundColor: getBadgeColor() }]}>
-          <Text style={styles.levelText}>
-            {getDifficultyLabel()}
-          </Text>
-        </View>
+        <DifficultyBadge />
       </View>
 
       {/* Right section: Timer and pause button */}
@@ -99,35 +63,6 @@ const styles = StyleSheet.create({
   scoreText: {
     fontSize: 14,
     fontWeight: '500',
-  },
-  levelLabel: {
-    fontSize: 10,
-    fontWeight: '700',
-    textTransform: 'uppercase',
-    marginBottom: 2,
-    color: '#666', // Subtle color for label
-    alignSelf: 'center',
-  },
-  levelBadge: {
-    paddingHorizontal: 10,
-    paddingVertical: 6,
-    borderRadius: 6, // Smaller corner radius to match other badges
-    marginTop: 2,
-    alignItems: 'center',
-    justifyContent: 'center',
-    minWidth: 80, // Increased to match timer badge
-    height: 36, // Fixed height to match other badges
-    elevation: 1, // Light shadow for Android
-    shadowColor: '#000', // Shadow for iOS
-    shadowOffset: { width: 0, height: 1 },
-    shadowOpacity: 0.2,
-    shadowRadius: 1,
-  },
-  levelText: {
-    fontSize: 16, // Match other badges
-    fontWeight: 'bold',
-    color: '#333333',
-    letterSpacing: 1,
   },
 });
